Use EdgeGridAuth to sign requests
This patch rips out our internal request signing tool and replaces it
with the EdgeGridAuth Python library. Users will need to install it
before they can use egcurl:

pip install edgegrid-python

I have maintained support for reading in request bodies with @filename
syntax (this is a curl feature).

Instead of using the Python "request" module I have constructed a
MockRequest to play the part silently. We make one of these, call on it
to produce the authorization header, then retrieve the header that was
produced.