Cleanup parsing of config file; automatic hostname selection
This patch cleans up how we parse the config file (now it's like DOM
parsing in XML) and cleans up the section selection and valication
logic.

There's an extra benefit change here. The hostname in your original
request is *ignored* in favor of the hostname specified in your config
section. This is better because the hostnames are aweful to copy/paste.
It's a bit of a maintenance burdon.