Warn about max-body of 8192
For a long time the Akamai Control Center mistakenly suggeted a max-body
value of 8192 when credentials were downloaded for EdgeGrid. This patch
adds a warning when we see that value in the selected config, because it
is very likely wrong. If it's wrong, POST requests above 8192KB length
will fail with signature mismatches.