Clean up behavior around `Expect` header
By default, curl will send `Expect: 100-continue` for all POST requests.
This is part of the HTTP spec, to allow origins to decide whether a
proposed content length is acceptable before the client actually sends
it. In the context of an API client, this is not really necessary or
helpful behavior. It's easier to just skip this step.

This patch does two things:
- Only sends `Expect` at all for POST requests.
- Adds code comments to explain why we do this.